Ticket #1548: Regenerate user certs with 10 years expiration and add regeneration process in script

+++ samples/ClientServer/data/cert/generate_user_certs.sh
@@ -17,6 +17,10 @@
 # specific language governing permissions and limitations
 # under the License.
 
+# This script is used to generate user root CA and user certificates,
+# with or without generation of new keys. Usage: 
+# - Generate user root CA + user certificates, with generation of new keys (default behaviour): "./generate_user_certs.sh"
+# - Generate user certificates with existing keys: "./generate_user_certs.sh --no-key-generation"
 
 set -e
 
@@ -27,37 +31,61 @@ CA_CERT_USR=user_cacert.pem
 
 DURATION=730
 
-# CA generation for users X509IdentityToken: generate key, generate self signed certificate
-# /!\ only for test as no pass phrase is embedeed
-openssl genrsa -out $CA_KEY_USR -aes-256-cbc 4096
-openssl req -config $CONF_CA_USR -new -x509 -key $CA_KEY_USR -out $CA_CERT_USR -days $DURATION
+# If no argument provided, default behaviour: regenerate user root CA + its crl, and user certificates, all with new keys.
+if [ $# == 0 ]; then
+    # CA generation for users X509IdentityToken: generate key, generate self signed certificate
+    # /!\ only for test as no pass phrase is embedeed
+    openssl genrsa -out $CA_KEY_USR -aes-256-cbc 4096
+    openssl req -config $CONF_CA_USR -new -x509 -key $CA_KEY_USR -out $CA_CERT_USR -days $DURATION
 
-# Generate an empty Certificate Revocation List, convert it to DER format
-openssl ca -config $CONF_CA_USR -gencrl -crldays $DURATION -out user_cacrl.pem
-openssl crl -in user_cacrl.pem -outform der -out user_cacrl.der
+    # Generate an empty Certificate Revocation List, convert it to DER format
+    openssl ca -config $CONF_CA_USR -gencrl -crldays $DURATION -out user_cacrl.pem
+    openssl crl -in user_cacrl.pem -outform der -out user_cacrl.der
 
-# Generate user X509IdentityToken 2048 key lengths, a new key pair
-openssl req -config $CONF_USR -reqexts user_cert -sha256 -nodes -newkey rsa:2048 -keyout user_2k_key.pem -out user_2k.csr
-openssl req -config $CONF_USR -reqexts user_cert -sha256 -nodes -newkey rsa:4096 -keyout user_4k_key.pem -out user_4k.csr
+    # Generate user X509IdentityToken 2048 key lengths, a new key pair
+    openssl req -config $CONF_USR -reqexts user_cert -sha256 -nodes -newkey rsa:2048 -keyout user_2k_key.pem -out user_2k.csr
+    openssl req -config $CONF_USR -reqexts user_cert -sha256 -nodes -newkey rsa:4096 -keyout user_4k_key.pem -out user_4k.csr
 
-# And sign them, for the next century
+    # Generate user encrypted private keys (these commands require the password).
+    echo "****** User private keys encryption ******"
+    openssl rsa -in user_2k_key.pem -aes-256-cbc -out encrypted_user_2k_key.pem
+    openssl rsa -in user_4k_key.pem -aes-256-cbc -out encrypted_user_4k_key.pem
+
+    # Remove the unencrypted keys
+    rm user*_key.pem
+
+    # Convert CA root into DER
+    openssl x509 -in $CA_CERT_USR -out user_cacert.der -outform der
+
+    # Print
+    echo -e "\nUser certificate Authority signed public key:"
+    hexdump -ve '/1 "%02x"' user_cacert.der
+    echo
+    echo -e "\nCRL of the CA:"
+    hexdump -ve '/1 "%02x"' user_cacrl.der
+    echo
+
+# If argument --no-key-generation provided: renew user certificates only (with existing keys)
+elif [ $# == 1 ] && [ $1 == "--no-key-generation" ]; then
+    # create csr for existing user keys
+    openssl req -new -config $CONF_USR -sha256 -key encrypted_user_2k_key.pem -reqexts user_cert -out user_2k.csr
+    openssl req -new -config $CONF_USR -sha256 -key encrypted_user_4k_key.pem -reqexts user_cert -out user_4k.csr
+
+# If invalid argument(s) provided: print usage and return
+else
+    echo "Please provide valid argument: --no-key-generation or no argument at all"
+    exit 1
+fi
+
+# And sign them, for the next two years
 openssl ca -batch -config $CONF_CA_USR -policy signing_policy -extensions user_signing_req -days $DURATION -in user_2k.csr -out user_2k_cert.pem
 openssl ca -batch -config $CONF_CA_USR -policy signing_policy -extensions user_signing_req -days $DURATION -in user_4k.csr -out user_4k_cert.pem
 
 # Remove the CSRs
 rm ./*.csr
 
-# Generate user encrypted private keys (these commands require the password).
-echo "****** User private keys encryption ******"
-openssl rsa -in user_2k_key.pem -aes-256-cbc -out encrypted_user_2k_key.pem
-openssl rsa -in user_4k_key.pem -aes-256-cbc -out encrypted_user_4k_key.pem
-
-# Remove the unencrypted keys
-rm user*_key.pem
-
-
 # Output application and user certificates in DER format
-for fradix in user_ca user_2k_ user_4k_; do
+for fradix in user_2k_ user_4k_; do
     openssl x509 -in ${fradix}cert.pem -out ${fradix}cert.der -outform der
 done
 
@@ -67,10 +95,4 @@ hexdump -ve '/1 "%02x"' user_2k_cert.der
 echo
 echo -e "\nUser 2k cert's thumbprint (SHA-1):"
 openssl x509 -noout -fingerprint -in user_2k_cert.pem
-echo
-echo -e "\nUser certificate Authority signed public key:"
-hexdump -ve '/1 "%02x"' user_cacert.der
-echo
-echo -e "\nCRL of the CA:"
-hexdump -ve '/1 "%02x"' user_cacrl.der
 echo
